services/normalizer/requirements.txt
redis==7.0.1
jmespath==1.0.1
clickhouse-driver==0.2.9
python-dotenv==1.0.1
